requirements/base.txt
# Sanic is a Flask-like Python 3.5+ web server that's written to go fast.
# https://github.com/channelcat/sanic
sanic==0.8.3

# A Sanic extension adding a decorator for CORS support. Based on flask-cors by Cory Dolphin.
# https://github.com/ashleysommer/sanic-cors
sanic-cors==0.9.7

# Requests is an elegant and simple HTTP library for Python, built for human beings.
# https://github.com/requests/requests
requests==2.19.1

# Decouple helps you to organize your settings so that you can change parameters without having to redeploy your app
# https://github.com/henriquebastos/python-decouple
python-decouple==3.1

# Cerberus provides powerful yet simple and lightweight data validation functionality out of the box and is designed to be easily extensible, allowing for custom validation
# http://docs.python-cerberus.org/en/stable/index.html
cerberus==1.2

# The Orator ORM provides a simple yet beautiful ActiveRecord implementation
# https://orator-orm.com/docs/0.9/installation.html
orator==0.9.7

# A Python slugify application that handles unicode.
# https://github.com/un33k/python-slugify
python-slugify==1.2.5

# Psycopg is the most popular PostgreSQL database adapter for the Python programming language.
# http://initd.org/psycopg/docs/
psycopg2-binary==2.7.5

# The Python interface to the Redis key-value store.
# https://github.com/andymccurdy/redis-py
redis==2.10.6

# Object serialization and deserialization, lightweight and fluffy.
# https://marshmallow.readthedocs.io/en/latest/install.html
marshmallow==2.15.4

# Pylint is a Python static code analysis tool which looks for programming errors, 
# helps enforcing a coding standard, sniffs for code smells and offers simple refactoring suggestions
# https://github.com/PyCQA/pylint/
pylint==2.3.0

# An(other) implementation of JSON Schema for Python
# https://github.com/Julian/jsonschema
jsonschema==2.6.0

# Raven is the official legacy Python client for Sentry, 
# officially supports Python 2.6–2.7 & 3.3–3.7, and runs on PyPy and Google App Engine.
# https://github.com/getsentry/raven-python
raven==6.9.0

# Python money class with optional CLDR-backed locale-aware formatting 
# and an extensible currency exchange solution.
# https://github.com/carlospalol/money
money==1.3.0

# Babel is an integrated collection of utilities that assist in internationalizing 
# and localizing Python applications, with an emphasis on web-based applications.
# http://babel.pocoo.org/en/latest/
Babel==2.6.0

Jinja2==2.10.3
